requirements.txt
flask~=2.2.3
apispec~=6.3.0
marshmallow~=3.19.0
requests~=2.29.0
pillow~=9.5.0
beautifulsoup4~=4.12.2
werkzeug~=2.3.1